add toggle-able prometheus shim layer
Adds an env var `RETRACED_ENABLE_PROMETHEUS` that will toggle on prometheus metrics and serve them on the default http server. (someday this should be a separate port though)

If Prometheus is enabled, the same monkit statsd metrics will still be reported. The way we do this today is a little janky, but I have ideas to clean it up if this gets past the "experimentation" phase.

Had to create our own copy-pasta'd versions of `instrument` and `instrumented` so we can override calls to `meter` and `timer` with our new maybe-prometheus-and-statsd implementations.