do not publish phantom IDs to raw_events
When the ingestFromQueue worker is called to handle events on the
unsaved_events topic, it generates an id and attempts to save that id
along with the event into the ingest_task table. If the event has
already been saved to the ingest_task table then there is nothing to be
done. In that case the generated id should not be published to
raw_events.