Fix connection pool contention (#427)
The normalizeEvent worker was using a single long-lived connection plus
four short-lived connections, i.e.  2 concurrent connections for most of the
lifetime of the job.  Five normalizeWorker jobs can be in-flight at
once, for a total of 10 concurrent connections, which was the default
limit on the pool.

This changes the normalizeEvent worker to use a single connection for
all queries. Also raise the default pool connection limit to 20.